requirements.txt
# Core dependencies
polars>=0.20.0
numpy>=1.24.0
pandas>=2.0.0  # Optional: for compatibility layer
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data I/O
pyyaml>=6.0
openpyxl>=3.1.0  # Excel support for Polars
pyarrow>=14.0.0  # Parquet support

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Statistics
statsmodels>=0.14.0

# Parallel processing
joblib>=1.3.0

# Logging
colorlog>=6.7.0

# Optional: SHAP for feature importance
shap>=0.43.0